backend/velora_db/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.10.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.0.1
PyJWT==2.10.1
requests==2.32.3
openai==1.99.1
python-multipart==0.0.20
reportlab==4.2.2
gTTS==2.5.4
